==> 01-Program-Years/2021/Project Code/dspg21RnD/src/04_eads_method/05_eads_nmf_fit.ipynb <==
# NMF Fitting

nmf_c = pd.read_pickle("/home/sbc8kw/git/dspg21RnD/data/dspg21RnD/nmf_bert_10.pkl")
nmf_c['n_topics'] = nmf_c.index
nmf_c.to_csv("/home/sbc8kw/git/dspg21RnD/data/dspg21RnD/nmf_bert_10.csv", index= False)

pwd

import pandas as pd
import numpy as np
import pickle
import matplotlib.pyplot as plt
import gensim
import time

from sklearn.decomposition import NMF, TruncatedSVD, LatentDirichletAllocation
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer

nmf_c = pd.read_pickle("../../data/dspg21RnD/nmf_Eads_10.pkl")


nmf_c['n_topics'] = nmf_c.index

nmf_c.to_csv("../../data/dspg21RnD/nmf_eads_10.csv", index= False)

def print_topics(model, vectorizer, top_n=10):
    for idx, topic in enumerate(model.components_):  # loop through each row of H.  idx = row index.  topic = actual row
        print("\nTopic %d:" % (idx))
        #print([(vectorizer.get_feature_names()[i], topic[i])  # printing out words corresponding to indices found in next line
                        #for i in topic.argsort()[:-top_n - 1:-1]])  # finding indices of top words in topic
            
        print_list = [(vectorizer.get_feature_names()[i], topic[i])  
                        for i in topic.argsort()[:-top_n - 1:-1]]
        for item in print_list:
            print(item)

#TODO:Input a dataframe that are ai related abstracts, need variables: final_frqwds_removed
abstracts = pd.read_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/Eads_AI_abstracts-KL.csv') 
abstracts2 = pd.read_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/Eads_AI_abstracts.csv')   

lim_docs = abstracts["final_frqwds_removed"]
len(lim_docs)

lim_docs2 = abstracts2["final_frqwds_removed"]
len(lim_docs2)

# input needed for LDA, NMF (all from Scikit-Learn) is one string per document (not a list of strings)

text1 = []

for token_list in lim_docs:
    text1.append(token_list)
    
    
    
text2 = []

for token_list in lim_docs2:
    text2.append(token_list)

# Create a TF-IDF document-term matrix for the AI corpus 

# TRY DIFFERENT PARAMETERS IN THE TF-IDF DOC-TERM MATRIX SET-UP
nmf_vectorizer1 = TfidfVectorizer(max_df=1.0, min_df=3, lowercase=True) #, max_features=int(len(lim_docs)/2))
nmf_vectorizer2 = TfidfVectorizer(max_df=1.0, min_df=3, lowercase=True) #, max_features=int(len(lim_docs)/2))

# by default TfidfVectorizer has l2 normalization for rows: 
# from Scikit Learn documentation: Each output row will have unit norm, either: * ‘l2’: Sum of squares of vector 
# elements is 1. The cosine similarity between two vectors is their dot product when l2 norm has been applied.

nmf_tf_idf1 = nmf_vectorizer1.fit_transform(text1)
nmf_tf_idf1.shape


nmf_tf_idf2 = nmf_vectorizer2.fit_transform(text2)
nmf_tf_idf2.shape

# Mean+2 SD (k=30)

# topic modeling with NMF
nmf_model1 = NMF(n_components=30, random_state=1)  # TRY DIFFERENT NUMBERS OF TOPICS
W1 = nmf_model1.fit_transform(nmf_tf_idf1)
H1 = nmf_model1.components_

print_topics(nmf_model1, nmf_vectorizer1, 15)

# Mean+2.5 SD(k=30)

#topic modeling with NMF

nmf_model2 = NMF(n_components=30, random_state=1)  # TRY DIFFERENT NUMBERS OF TOPICS
W2 = nmf_model2.fit_transform(nmf_tf_idf2)
H2 = nmf_model2.components_
#print_topics(nmf_model, nmf_vectorizer, 10)

print_topics(nmf_model2, nmf_vectorizer2, 15)
